+++ core/src/main/scala/mongo4s/queries/FindQuery.scala
@@ -1,11 +1,14 @@
 package mongo4s.queries
 
+import scala.NamedTuple.AnyNamedTuple
 import scala.concurrent.duration.FiniteDuration
+import scala.quoted.*
 
 import org.bson.BsonDocument
 
-import mongo4s.Streamable
+import mongo4s.bson.{BsonDocumentDecoder, FieldNaming}
 import mongo4s.operations.{Filter, Projection, Sort}
+import mongo4s.{SelectMacro, Streamable}
 
 trait FindQuery[F[*], S[*], A]:
   def filter(filter: Filter[A]): FindQuery[F, S, A]
@@ -25,3 +28,8 @@ trait FindQuery[F[*], S[*], A]:
   def stream(using Streamable[S, A]): S[A]
 
   def attempting: DecodeAttempts[F, S, A]
+
+  def selecting[B](projection: Projection[A], decoder: FieldNaming => BsonDocumentDecoder[B]): SelectQuery[F, S, B]
+
+  inline def selectAs[K <: AnyNamedTuple]: SelectQuery[F, S, K] =
+    ${ SelectMacro.impl[F, S, A, K]('this) }